crypto: add certificate manager to find certificates either locally or in k8s
Map keys "tls.key" and "tls.crt" are used for the private key and the
certificate, respectively, when creating a TLS secret like so:

$ kubectl create secret tls test-tls --key="file1" --cert="file2"

Change item names we look for to be compatible with this.

Add support for explicit overrides for the default item names. If
given, these must be found.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>
Signed-off-by: André Martins <andre@cilium.io>